requirements.txt
streamlit==1.52.1
rdkit==2025.9.1
xgboost==3.1.2
joblib==1.5.2
pandas==2.3.3
numpy==2.2.6
Pillow
